FIX : #63 convert sideBar to sidebar, import path error
