+++ src/views/chatroom/chatroom.tsx
@@ -5,12 +5,12 @@ import { useState, useRef } from "react";
 import { useParams } from "react-router-dom";
 import Header from "../../components/header/header";
 import ModalHandler from "../../components/modal/modalhandler";
-import SideBar from "../../components/sideBar/sideBar";
-import { sidebarProperty } from "../../components/sideBar/sideBarType";
+import { sidebarProperty } from "../../components/sidebar/sidebarType";
 import ChatMessage from "./message/message";
 import { ioChannel } from "../../socket/socket";
 import "./chatroom.scss";
 import GameModalListener from "../../components/modal/gameModalListener";
+import ChatroomSidebar from "../../components/sidebar/chatroomSidebar";
 
 // 서버로부터 받아서 message state 에 넣을 때 들어가는 형태
 type Message = {
@@ -48,6 +48,10 @@ const Chatroom = () => {
       };
       setMessages([...messages, newMessage]);
     }
+    //post joinChannel -> data 없이 
+    //결과값 따라서 false면 아래 alert 가져와서 하고
+    //true면 sidebar socket
+    //useeefect 끝날때 return () => {} -> unsubscribe 
     ioChannel.emit("joinChannel", id);
     ioChannel.on("joinRefused", () => {
       window.location.href = "http://localhost:3000/main";
@@ -70,6 +74,7 @@ const Chatroom = () => {
 	}
 
   const leaveChannel = () => {
+    //post 
     ioChannel.emit("leaveChannel", id);
     window.location.href = "http://localhost:3000/main";
   }
@@ -78,8 +83,7 @@ const Chatroom = () => {
     <>
       <Header isLoggedIn={true} />
       <div className="page">
-        <SideBar
-          title={sidebarProperty.chatMemberList}
+        <ChatroomSidebar
           roomId={id}
           modalHandler={modalHandler}
         />
