+++ src/views/game/game.tsx
@@ -5,9 +5,7 @@ import { io } from "../../socket/socket";
 import { GameInfo } from "./gameType";
 import Header from "../../components/header/header";
 import Pong from "./pong/pong";
-import SideBar from "../../components/sideBar/sideBar";
 import "./game.scss";
-import { sidebarProperty } from "../../components/sideBar/sideBarType";
 import ModalHandler from "../../components/modal/modalhandler";
 
 function Game() {
@@ -56,11 +54,6 @@ function Game() {
     <>
       <Header isLoggedIn={true} />
       <div className="page">
-        <SideBar
-          title={sidebarProperty.observerList}
-          roomId={Number(id)}
-          modalHandler={modalHandler}
-        />
         {gameInfo ? (
           <div className="game-wrap">
             <div className="game-scoreboard">
